backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.9.5
celery==5.4.0
redis==5.0.7
mysqlclient==2.2.4
drf-spectacular==0.27.2
gunicorn==22.0.0
